feat: inclui solução do simulated annealing

+++ src/SA.h
@@ -8,11 +8,11 @@
 #include <vector>
 
 float SA(int n, std::vector<int> &s, float **d, 
-	     float alpha, float SAmax, double temp_inicial, 
+	     float alpha, int SAmax, double temp_inicial,
 	     double temp_final);
 
 float temperaturaInicial(int n, std::vector<int> &s, float **d, 
-	                     float beta, float gama, float SAmax, 
+	                     float beta, float gama, float SAmax,
 	                     double temp_inicial);
 
 #endif //PCV_SA_H
